PySide compatibility
Adds compatibility for the PySide Qt API. This introduces a few niggles
due to known bugs in PySide or missing/difference elements. Briefly -

- there appears to be a problem with singleShot timers losing the reference to the C++ object. I've resolved
this by re-creating the timers on each shot, but this may introduce overhead. An alternative is to revert
to the previous approach and catch the RuntimeError that results.
- `hasPendingEvents` always returns True and hangs the tests. There are some known problem with `hasPendingEvents`
always returning True in certain versions of Qt (±PySide) but not clear if that is the fault here. Either way, running
that function once and then continuing passes the test.
- qWait test function is not available in PySide. It's been added in` qt.py`